Move enums from types file
